requirements.txt
google.generativeai
langchain
faiss-cpu
openai
langchain-community
tiktoken